demos-no-framework/pyproject.toml
[project]
name = "demos-no-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "mcp>=1.17.0",
    "pip>=25.3",
    "python-dotenv>=1.1.1",
    "questionary>=2.1.1",
]
